Separate tests for gmtinfo and grdinfo
