+++ pygmt/tests/test_info.py
@@ -6,9 +6,8 @@ import os
 import numpy as np
 import pytest
 
-from .. import info, grdinfo
+from .. import info
 from ..exceptions import GMTInvalidInput
-from ..datasets import load_earth_relief
 
 TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "data")
 POINTS_DATA = os.path.join(TEST_DATA_DIR, "points.txt")
@@ -53,22 +52,3 @@ def test_info_fails():
         info(fname=21)
     with pytest.raises(GMTInvalidInput):
         info(fname=np.arange(20))
-
-
-def test_grdinfo():
-    "Make sure grd info works as expected"
-    grid = load_earth_relief()
-    result = grdinfo(grid, L=0, C="n")
-    assert result.strip() == "-180 180 -90 90 -8592.14453125 5558.79248047 1 1 361 181"
-
-
-def test_grdinfo_file():
-    "Test grdinfo with file input"
-    result = grdinfo("@earth_relief_60m", L=0, C="n")
-    assert result.strip() == "-180 180 -90 90 -8592.14465255 5558.79248047 1 1 361 181"
-
-
-def test_grdinfo_fails():
-    "Check that grdinfo fails correctly"
-    with pytest.raises(GMTInvalidInput):
-        grdinfo(np.arange(10).reshape((5, 2)))
